Fix right panel background extending beyond the truncation symbol
In side-by-side mode, if `background_color_extends_to_terminal_width`
is set, the left panel color is extended via spaces, but the right
one via an ANSI sequence which instructs the terminal emulator to
fill the background color rightwards.

However due to the padding after the right panel the color will
extend beyond a possible truncation symbol:

"TEXT#AB→#", not "TEXT#AB→ " with "#" being the green background.

This is fixed by using the same logic as in the left panel for the
right panel.